requirements.txt
fuzzywuzzy[speedup]
numpy
openai
python-dotenv
pyzmq
sounddevice
speechrecognition
